refactor(04-01): git mv ping_generator.py -> time_generator.py (history-preserving)
- pure module move; class names unchanged in this commit (rename edits follow)
- backtest_trading_system import path repointed to keep the suite green
