+++ itrader/trading_system/backtest_trading_system.py
@@ -11,7 +11,7 @@ from itrader.order_handler.order_handler import OrderHandler
 from itrader.order_handler.storage import OrderStorageFactory
 from itrader.portfolio_handler.portfolio_handler import PortfolioHandler
 from itrader.execution_handler.execution_handler import ExecutionHandler
-from itrader.trading_system.simulation.ping_generator import PingGenerator
+from itrader.trading_system.simulation.time_generator import PingGenerator
 from itrader.universe.dynamic import DynamicUniverse
 from itrader.reporting.statistics import StatisticsReporting
